// modules/user-profile/types/branch.ts
export type Branch = {
  id: string;
  parent_id: null;
  name: string;
  phone: string;
  phone_code: string;
  email: string;
  latitude: string;
  longitude: string;
  country_id: string;
  state_id: string;
  city_id: string;
  country_name: string;
  state_name: string;
  city_name: string;
};